# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FST Digital API Documentation",
        "description": "API documentation for FST Digital Backend",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/modules/{id}/name": {
            "get": {
                "tags": [
                    "Modules"
                ],
                "summary": "Get module name by ID",
                "operationId": "1684591213d2f7b22ccd1c11dd2d6025",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Module ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Module name",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "module_intitule": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Module not found"
                    }
                }
            }
        },
        "/api/professeurs/by-departement": {
            "get": {
                "tags": [
                    "Professeurs"
                ],
                "summary": "Get professors by department",
                "operationId": "0c3470201533a3cdbbfc073f76e67411",
                "parameters": [
                    {
                        "name": "departement",
                        "in": "query",
                        "description": "Department name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of professors in the department",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "nom": {
                                                "type": "string"
                                            },
                                            "prenom": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "departement": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/students/by-exam/{examId}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get students by exam ID",
                "operationId": "603c9fa35529f32f1a6f327495f3ccd0",
                "parameters": [
                    {
                        "name": "examId",
                        "in": "path",
                        "description": "Exam ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of students for the exam",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "nom": {
                                                "type": "string"
                                            },
                                            "prenom": {
                                                "type": "string"
                                            },
                                            "numero_etudiant": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/superviseurs/by-departement": {
            "get": {
                "tags": [
                    "Superviseurs"
                ],
                "summary": "Get superviseurs by department",
                "operationId": "5aef3b306e8588d7211b8c4e23a9b605",
                "parameters": [
                    {
                        "name": "departement",
                        "in": "query",
                        "description": "Department name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of superviseurs in the department",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "nom": {
                                                        "type": "string"
                                                    },
                                                    "prenom": {
                                                        "type": "string"
                                                    },
                                                    "poste": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
...
}